third_party/WebKit/LayoutTests/fast/multicol/vertical-lr/balancing/balance-trailing-border.html

<!DOCTYPE html>
<script src="../../../../resources/js-test.js"></script>
<p>Both words below should be in the the right column.</p>
<div id="console"></div>
<div id="mc" style="-webkit-writing-mode:vertical-lr; -webkit-columns:2; columns:2; orphans:1; widows:1; -webkit-column-rule:1px solid; column-rule:1px solid; background:olive;">
    <div style="width:110px"></div>
    <div style="border:2px solid black; height:175px;">
        <div style="margin:10px 0; border:2px solid blue;">
            <span style="line-height:80px;">right<br>column</span>
        </div>
    </div>
</div>
<script>
    description("Tests that the bottom border at a column boundary is recorded properly for balancing");
    shouldBe("mc.offsetWidth", "168");
</script>
